% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{awesomelab4}
\alias{awesomelab4}
\alias{awesomelab4-package}
\title{awesomelab4: Creating a linear regression package.}
\description{
TODO
}
\section{Functions}{

\code{placeholder()} returns scalars a and b.
}

\author{
Martin Svensson and Laura Julia Melis.

Course in Advanced Programming in R.

Departmenf of Computer and Information Science, Linköping University.
}
