% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class.}
\description{
A Reference Class.
}
\section{Fields}{

\describe{
\item{\code{formula}}{TODO}

\item{\code{data}}{TODO}

\item{\code{regression_coefficients}}{TODO}

\item{\code{fitted_values}}{TODO}

\item{\code{residuals}}{TODO}

\item{\code{degrees_of_freedom}}{TODO}

\item{\code{residual_variance}}{TODO}

\item{\code{variance_of_coefficients}}{TODO}

\item{\code{std_error}}{TODO}

\item{\code{t_values}}{TODO}

\item{\code{p_values}}{TODO}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{initialize(formula, data)}}{Initialize function calculates all values needed from formula and data.}

\item{\code{pred()}}{Returns the predictions.}

\item{\code{resid()}}{Returns the residuals.}

\item{\code{summary()}}{Prints a summary for the calculated data.}
}}

\examples{
data("iris")
linreg$new(formula=Petal.Length~Species, data=iris)

}
\references{
#TODO#
}
