% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Multiple Regression Model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{an object of class "\code{formula}" describing the model sctructure.}

\item{data}{an object of class "data frame" containing the variables in the model.}
}
\value{
\code{linreg} returns an object of class "\code{}"
}
\description{
Multiple Regression Model works by... #TODO#
}
\examples{
linreg(formula=Petal.Length~Species, data=iris)

}
\references{
#TODO#
}
