% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{awesomelab4}
\alias{awesomelab4}
\alias{awesomelab4-package}
\title{awesomelab4: A linear regression package.}
\description{
The awesomelab4 package contains an object with class \code{linreg}, created with the RC object oriented system.
This object is helpful to handle linear regression models as it provides multiple methods such as \code{resid()} and
\code{coef()} to obtain the vector of residuals and the regressions coefficients of a given regression model, respectively.
}
\section{Class object}{

\code{linreg()} calculates various statistics related to the linear regression model and contains multiple methods to obtain the regression coefficients or the predicted values, among others.
}

\author{
Martin Svensson and Laura Julia Melis.

Course in Advanced Programming in R.

Departmenf of Computer and Information Science, Linköping University.
}
