% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class.}
\description{
\code{linreg()} calculates a number of statistics (see section \code{"Fields"}) by default given both arguments: formula and data. 
It also stores diverse functions implemented using RC object oriented programming. Find more information regarding these functions on section \code{"Methods"}.
}
\section{Fields}{

\describe{
\item{\code{formula}}{an object of class "formula".}

\item{\code{data}}{a data frame.}

\item{\code{regression_coefficients}}{the regression coefficients of the model calculated using the QR decomposition.}

\item{\code{fitted_values}}{the fitted values.}

\item{\code{residuals}}{the residuals. It is the difference between the real values and the fitted values.}

\item{\code{degrees_of_freedom}}{the degrees of freedom: the number of observations minus the number of parameters in the model.}

\item{\code{residual_variance}}{the residual variance.}

\item{\code{variance_of_coefficients}}{the variance of the regression coefficients.}

\item{\code{std_error}}{the standard error of the regression coefficients. It is the square root of the variance.}

\item{\code{t_values}}{the t-values for each coefficient.}

\item{\code{p_values}}{the p-values for each coefficient.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{initialize(formula, data)}}{Initialize function calculates all values needed from formula and data.}

\item{\code{pred()}}{Returns the predictions.}

\item{\code{print()}}{Same as show, but allows for linreg$print(). Will break print(linreg) that show allows, bad test case!}

\item{\code{resid()}}{Returns the residuals.}

\item{\code{summary()}}{Prints a summary for the calculated data.}
}}

\examples{
data("iris")
example <- linreg$new(formula=Petal.Length~Species, data=iris)

}
\references{
Reference Classes: \url{http://adv-r.had.co.nz/R5.html}

QR decomposition: \url{https://en.wikipedia.org/wiki/QR_decomposition}

Vignettes: \url{http://r-pkgs.had.co.nz/vignettes.html}
}
